Use floating point in pitch shifter again

+++ ports/pico/apps/sample_player/timestretched/pitch_shifter.h
@@ -21,7 +21,7 @@ struct PitchShifter : SampleReader {
   // Reader interface
   uint32_t read_samples(int16_t *out);
 
-  void set_speed(const double speed) {
+  void set_speed(const float speed) {
     if (speed < 0.2) {
       this->speed = 0.2;
     } else if (speed > 1.8) {
@@ -35,11 +35,11 @@ private:
   uint32_t read_resampled(int16_t *out);
   void shift_interpolation_samples(int16_t sample);
 
-  double speed;
+  float speed;
   int16_t interpolation_samples[4];
   uint32_t source_index;
-  uint32_t position;
-  int remainder;
+  float position;
+  float remainder;
   SampleReader &sample_reader;
   BufferedReader buffered_reader;
 };
